Do not specify "./" as the path on the server
If worked before, it stopped working at some point.
We are likely to make it more robust on dandi-cli side:
- https://github.com/dandi/dandi-cli/pull/1454

but I think it would be better to just not specify such
way in the example. You might want to elaborate that example
with actually pointing to a specific folder instead